Roave/BackwardCompatibilityCheck#767: Rename ClassBased/EnumCaseAdded to EnumBased/CaseAdded use in CompareClasses s
