+++ bin/roave-backward-compatibility-check.php
@@ -14,6 +14,7 @@ use Roave\BackwardCompatibility\Command;
 use Roave\BackwardCompatibility\CompareClasses;
 use Roave\BackwardCompatibility\DetectChanges\BCBreak\ClassBased;
 use Roave\BackwardCompatibility\DetectChanges\BCBreak\ClassConstantBased;
+use Roave\BackwardCompatibility\DetectChanges\BCBreak\EnumBased\CaseAdded;
 use Roave\BackwardCompatibility\DetectChanges\BCBreak\FunctionBased;
 use Roave\BackwardCompatibility\DetectChanges\BCBreak\InterfaceBased;
 use Roave\BackwardCompatibility\DetectChanges\BCBreak\MethodBased;
@@ -312,6 +313,7 @@ use function file_exists;
                     ),
                 )),
             ))),
+            new CaseAdded(),
         ),
     );
