+++ src/DetectChanges/BCBreak/EnumBased/EnumBased.php
@@ -1,0 +1,11 @@
+<?php
+
+namespace Roave\BackwardCompatibility\DetectChanges\BCBreak\EnumBased;
+
+use Roave\BackwardCompatibility\Changes;
+use Roave\BetterReflection\Reflection\ReflectionClass;
+
+interface EnumBased
+{
+    public function __invoke(ReflectionClass $fromInterface, ReflectionClass $toInterface): Changes;
+}
